fixed supprimer methods

+++ src/controleur/ControleurMesBiens.java
@@ -14,6 +14,7 @@ import java.util.LinkedList;
 import java.util.List;
 
 import javax.swing.JButton;
+import javax.swing.JOptionPane;
 
 import modeleDAO.BienDAO;
 import modeleDAO.ImmeubleDAO;
@@ -55,7 +56,9 @@ public class ControleurMesBiens implements ActionListener {
             this.bien = new LinkedList<>();
             
             BienDAO bien = new BienDAO();
-            ResultSet rs = bien.getAllBiens();
+            ImmeubleDAO immeuble = new ImmeubleDAO();
+            ResultSet rs = bien.getBiensFromOneImmeuble(this.vue.getIdImmeuble());
+			ResultSet rs2 = immeuble.getInfoImmeuble(this.vue.getIdImmeuble());
             while(rs.next()) {
                 if (rs.getString(2).equals("L")) {
                     this.bien.add(new Logement(rs.getDate(6), rs.getString(1), rs.getInt(3), rs.getInt(5), rs.getFloat(4)));
@@ -80,9 +83,18 @@ public class ControleurMesBiens implements ActionListener {
 				error.printStackTrace();
 			}
 		} else if (b.getText() == "Supprimer") {
-			BienDAO bien = new BienDAO();
-			int rs = bien.supprimerBien(this.bien.get(this.vue.getLigneChoisi()).getId_bien());
-			this.Update();
+			String[] options = {"Suppimer","Annuler"};
+			JOptionPane pane = new JOptionPane();
+			@SuppressWarnings("static-access")
+			int resultat=pane.showOptionDialog(this.vue, 
+					"Tous les locations associées à ce bien seront également supprimées.",
+					"Attention", 
+					JOptionPane.YES_NO_OPTION,JOptionPane.QUESTION_MESSAGE,null, options, options[1]);
+			if (resultat==JOptionPane.YES_OPTION) {
+				BienDAO bien = new BienDAO();
+				bien.supprimerBien(this.bien.get(this.vue.getLigneChoisi()).getId_bien());
+				this.Update();
+			}
 		}
 		
 	}
